Replace boost with capnp in RPC
Summary:
Converts the RPC stack to use Cap'n Proto for serialization instead of
boost. There are still some traces of boost in other places in the code,
but most of it is removed. A future diff should cleanup boost for good.

The RPC API is now changed to be more flexible with regards to how
serialize data. This makes the simplest cases a bit more verbose, but
allows complex serialization code to be correctly written instead of
relying on hacks. (For reference, look for the old serialization of
`PullRpc` which had a nasty pointer hacks to inject accessors in
`TypedValue`.)

Since RPC messages were uselessly modeled via inheritance of Message
base class, that class is now removed. Furthermore, that approach
doesn't really work with Cap'n Proto. Instead, each message type is
required to have some type information. This can be automated, so
`define-rpc` has been added to LCP, which hopefully simplifies defining
new RPC request and response messages.

Specify Cap'n Proto schema ID in cmake

This preserves Cap'n Proto generated typeIds across multiple generations
of capnp schemas through LCP. It is imperative that typeId stays the
same to ensure that different compilations of Memgraph may communicate
via RPC in a distributed cluster.

Use CLOS for meta information on C++ types in LCP

Since some structure slots and functions have started to repeat
themselves, it makes sense to model C++ meta information via Common Lisp
Object System.

Depends on D1391

Reviewers: buda, dgleich, mferencevic, mtomic, mculinovic, msantl

Reviewed By: msantl

Subscribers: pullbot

Differential Revision: https://phabricator.memgraph.io/D1407

+++ src/distributed/bfs_rpc_clients.hpp
@@ -19,39 +19,39 @@ class BfsRpcClients {
                 distributed::BfsSubcursorStorage *subcursor_storage,
                 distributed::RpcWorkerClients *clients);
 
-  std::unordered_map<int, int64_t> CreateBfsSubcursors(
+  std::unordered_map<int16_t, int64_t> CreateBfsSubcursors(
       tx::TransactionId tx_id, query::EdgeAtom::Direction direction,
       const std::vector<storage::EdgeType> &edge_types,
       query::GraphView graph_view);
 
   void RegisterSubcursors(
-      const std::unordered_map<int, int64_t> &subcursor_ids);
+      const std::unordered_map<int16_t, int64_t> &subcursor_ids);
 
   void RemoveBfsSubcursors(
-      const std::unordered_map<int, int64_t> &subcursor_ids);
+      const std::unordered_map<int16_t, int64_t> &subcursor_ids);
 
   std::experimental::optional<VertexAccessor> Pull(
-      int worker_id, int64_t subcursor_id, database::GraphDbAccessor *dba);
+      int16_t worker_id, int64_t subcursor_id, database::GraphDbAccessor *dba);
 
-  bool ExpandLevel(const std::unordered_map<int, int64_t> &subcursor_ids);
+  bool ExpandLevel(const std::unordered_map<int16_t, int64_t> &subcursor_ids);
 
-  void SetSource(const std::unordered_map<int, int64_t> &subcursor_ids,
+  void SetSource(const std::unordered_map<int16_t, int64_t> &subcursor_ids,
                  storage::VertexAddress source_address);
 
   bool ExpandToRemoteVertex(
-      const std::unordered_map<int, int64_t> &subcursor_ids, EdgeAccessor edge,
-      VertexAccessor vertex);
+      const std::unordered_map<int16_t, int64_t> &subcursor_ids,
+      EdgeAccessor edge, VertexAccessor vertex);
 
   PathSegment ReconstructPath(
-      const std::unordered_map<int, int64_t> &subcursor_ids,
+      const std::unordered_map<int16_t, int64_t> &subcursor_ids,
       storage::EdgeAddress edge, database::GraphDbAccessor *dba);
 
   PathSegment ReconstructPath(
-      const std::unordered_map<int, int64_t> &subcursor_ids,
+      const std::unordered_map<int16_t, int64_t> &subcursor_ids,
       storage::VertexAddress vertex, database::GraphDbAccessor *dba);
 
-  void PrepareForExpand(const std::unordered_map<int, int64_t> &subcursor_ids,
-                        bool clear);
+  void PrepareForExpand(
+      const std::unordered_map<int16_t, int64_t> &subcursor_ids, bool clear);
 
  private:
   database::GraphDb *db_;
